refactor: 抽出jmList
