+++ src/common/jmUtils.js
@@ -139,7 +139,7 @@ class jmUtils {
         let target = evt.target || evt.srcElement;
         if(touches) {
             evt = touches[0];//兼容touch事件
-            evt.target = target;
+            if(!evt.target) evt.target = target;
             isTouch = true;
         }
         let px = evt.pageX || evt.x;
@@ -558,228 +558,5 @@ class jmUtils {
         }
         return r;
     }
-
-    /**
-     * 定义一个属性
-     *
-     * @method createProperty
-     * @param {object} instance 被添加属性的对象
-     * @param {string} name 设置属性的名称
-     * @param {any} value 属性的值，可选, 如果直接指定为descriptor，也可以。可以传递get,set，按defineProperty第三个参数做参考
-     * @retunr {any} 当前属性的值
-     */
-    static createProperty(instance, name, value) {
-        var descriptor = {
-            configurable: false, //当且仅当该属性的 configurable 为 true 时，该属性描述符才能够被改变，同时该属性也能从对应的对象上被删除。默认为 false。
-            enumerable: true,
-            //valueMaps: {},
-            name: name,
-            defaultValue: value
-        };
-        if(value) {
-            if(!(value instanceof jmUtils.list) && (typeof value.set == 'function' || typeof value.get == 'function')) {
-                descriptor.set = value.set;
-                descriptor.get = value.get;
-                //descriptor.value = value.value; //赋值不能跟get/set同时存在
-            }
-            else if(typeof value == 'object' && ('value' in value || 'configurable' in value || 'enumerable' in value || 'writable' in value)) {
-                descriptor.value = value.value;
-                descriptor.configurable = typeof value.configurable == 'boolean'?value.configurable:true;
-                descriptor.enumerable = typeof value.enumerable == 'boolean'?value.enumerable:true;
-                descriptor.writable = typeof value.writable == 'boolean'?value.writable:true;
-            }
-            /*else {
-                descriptor.value = value;
-                descriptor.enumerable = true; //能否for in 默认false
-                descriptor.writable = true;
-            }*/
-        }
-        /*else {
-            descriptor.enumerable = true;
-            descriptor.writable = true;
-        }*/
-        //如果没有指定一些必要属性，则采用get/set方式初始化
-        if(!('get' in descriptor || 'set' in descriptor || 'value' in descriptor || 'writable' in descriptor)) {
-            descriptor.get = function() {
-                //如果是canvas对象，则直接返回
-                if(this.canvas && ('width' == descriptor.name || 'height' == descriptor.name )) {
-                    return this.canvas[descriptor.name];
-                }
-                this.__properties = this.__properties||{};
-                return this.__properties[descriptor.name];
-            };
-            descriptor.set = function(value) {
-                //如果是canvas 则修改其属性
-                if(this.canvas && ('width' == descriptor.name || 'height' == descriptor.name )) {
-                    this.canvas[descriptor.name] = value;
-                }
-                else {
-                    this.__properties = this.__properties||{};
-                    //var oldvalue = this.__properties[descriptor.name];
-                    this.__properties[descriptor.name] = value;
-                }
-                this.needUpdate = true;
-            };
-        }
-        //给对象定义一个属性
-        return Object.defineProperty(instance, name, descriptor);
-    }
 }
-
-/**
- * 自定义集合
- * 
- * @class list
- * @for jmUtils
- * @param {array} [arr] 数组，可转为当前list元素
- */
-jmUtils.list = class extends Array {
-    option = {}; //选项
-    constructor(...arg) {
-        let ps = [];
-        if(arg && arg.length && Array.isArray(arg[0])) {
-            for(let i=0; i< arg[0].length; i++) ps.push(arg[0][i]);
-            super(...ps);
-        }
-        else {
-            super();
-        }
-    }
-    /**
-     * 往集合中添加对象
-     *
-     * @method add
-     * @for list
-     * @param {any} obj 往集合中添加的对象
-     */
-    add(obj) {        
-        if(obj && Array.isArray(obj)) {
-            for(let i=0; i < obj.length; i++) {
-                if(!this.includes(obj[i])) this.push(obj[i]);
-            } 
-            return obj;           
-        }
-        if(typeof obj == 'object' && this.includes(obj)) return obj;
-        this.push(obj);
-        return obj;
-    }
-
-    /**
-     * 从集合中移除指定对象
-     * 
-     * @method remove
-     * @for list
-     * @param {any} obj 将移除的对象
-     */
-    remove(obj) {
-        for(let i = this.length -1; i>=0; i--) {            
-            if(this[i] == obj) {
-                this.removeAt(i);
-            }
-        }
-    }
-
-    /**
-     * 按索引移除对象
-     * 
-     * @method removeAt
-     * @for list
-     * @param {integer} index 移除对象的索引
-     */
-    removeAt(index) {
-        if(this.length > index) {
-            let obj = this[index];
-            this.splice(index,1);
-            if(this.option.removeHandler)  this.option.removeHandler.call(this, obj, index);
-        }
-    }
-
-    /**
-     * 判断是否包含某个对象
-     * 
-     * @method contain
-     * @for list
-     * @param {any} obj 判断当前集合中是否包含此对象
-     */
-    contain(obj) {
-        return this.includes(obj);
-    }
-
-    /**
-     * 从集合中获取某个对象
-     * 
-     * @method get
-     * @for list
-     * @param {integer/function} index 如果为整型则表示为获取此索引的对象，如果为function为则通过此委托获取对象
-     * @return {any} 集合中的对象
-     */
-    get(index) {
-        if(typeof index == 'function') {
-            return this.find(index);
-        }
-        else {
-            return this[index];
-        }        
-    }
-
-    /**
-     * 遍历当前集合 
-     *
-     * @method each
-     * @for list
-     * @param {function} cb 遍历当前集合的委托
-     * @param {boolean} inverse 是否按逆序遍历
-     */
-    each(cb, inverse) {
-        if(cb && typeof cb == 'function') {
-            //如果按倒序循环
-            if(inverse) {
-                for(let i = this.length - 1;i>=0; i--) {
-                    let r = cb.call(this, i, this[i]);
-                    if(r === false) break;
-                }
-            }
-            else {
-                let len = this.length;
-               for(let i  = 0; i < len;i++) {
-                    let r = cb.call(this, i, this[i]);
-                    if(r === false) break;
-                } 
-            }            
-        }        
-    }
-
-    /**
-     * 获取当前集合对象个数
-     *
-     * @method count
-     * @param {function} [handler] 检查对象是否符合计算的条件
-     * @for list
-     * @return {integer} 当前集合的个数
-     */
-    count(handler) {
-        if(handler && typeof handler == 'function') {
-            let count = 0;
-            let len = this.length;
-            for(let i  = 0; i<len;i++) {
-                if(handler(this[i])) {
-                    count ++;
-                }
-            } 
-            return count;
-        }
-        return this.length;
-    }
-
-    /**
-     * 清空当前集合
-     *
-     * @method clear
-     * @for list
-     */
-    clear() {
-        this.splice(0, this.length);
-    }
-}
-
 export default jmUtils;
